IndexPairを専用のヘッダーに移動
