+++ TowerGame.vcxproj
@@ -210,6 +210,7 @@
     <ClInclude Include="FireFly.h" />
     <ClInclude Include="FPS.h" />
     <ClInclude Include="Game.h" />
+    <ClInclude Include="IndexPair.h" />
     <ClInclude Include="InputManager.h" />
     <ClInclude Include="LeftAttackState.h" />
     <ClInclude Include="LeftMoveState.h" />
